#31: PlaceMapActivity.kt 파일의 패키지 변경
